feat: add NextAuth configuration and refactor authentication imports
