+++ app/[locale]/dashboard/page.tsx
@@ -4,7 +4,5 @@ import { DashboardContent } from "./content";
 export default async function DashboardPage() {
   const session = await auth();
 
-  console.log(session)
-
   return <DashboardContent session={session} />;
 }
