+++ auth.config.ts
@@ -1,5 +1,5 @@
 import type { NextAuthConfig } from "next-auth";
-import { credentialsProvider } from "./credentials";
+import { credentialsProvider } from "./lib/auth/credentials";
 
 // Notice this is only an object, not a full Auth.js instance
 export default {
